Moved title component to text components with subtitle
